Pulled logging into its own module.
