+++ src/gantry/cli/_main.py
@@ -6,7 +6,20 @@ import click
 from . import build, configure, schemas
 
 from .. import __version__
-from .._types import Path, ProgramOptions, LOGGER_NAME
+from .._types import Path, ProgramOptions
+from ..logging import LOGGER_NAME
+
+
+def _configure_logger() -> None:
+    ch = logging.StreamHandler(sys.stdout)
+    ch.setLevel(logging.DEBUG)
+
+    formatter = logging.Formatter('%(levelname)s :: %(message)s')
+    ch.setFormatter(formatter)
+
+    logger = logging.getLogger(LOGGER_NAME)
+    logger.setLevel(logging.DEBUG)
+    logger.addHandler(ch)
 
 
 @click.group()
@@ -21,15 +34,7 @@ from .._types import Path, ProgramOptions, LOGGER_NAME
 @click.pass_context
 def main(ctx: click.Context, services_path: Path):
     '''A container orchestrator for homelabs.'''
-    ch = logging.StreamHandler(sys.stdout)
-    ch.setLevel(logging.DEBUG)
-
-    formatter = logging.Formatter('%(levelname)s :: %(message)s')
-    ch.setFormatter(formatter)
-
-    logger = logging.getLogger(LOGGER_NAME)
-    logger.setLevel(logging.DEBUG)
-    logger.addHandler(ch)
+    _configure_logger()
 
     ctx.obj = ProgramOptions(services_path=services_path)
