+++ src/gantry/targets/__init__.py
@@ -1,3 +1,3 @@
 from ._common import Target, copy_services_resources
 from .compose import ComposeTarget
-from .image import BuildDockerImages
+from .image import ImageTarget
